requirements.txt
amqp==5.2.0
anyio==4.3.0
APScheduler==3.10.4
asgiref==3.6.0
async-timeout==4.0.3
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
Babel==2.14.0
billiard==4.2.0
bizdays==1.0.13
bleach==6.1.0
boto3==1.40.0
botocore==1.40.0
Brotli==1.1.0
cachetools==5.3.3
celery==5.3.6
certifi==2024.2.2
cffi==1.16.0
channels-redis==4.2.1
channels==4.0.0
channelsmultiplexer==0.0.3
chardet==5.2.0
charset-normalizer==3.3.2
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
click==8.1.7
constantly==23.10.4
cryptography==43.0.3
daphne==4.1.0
dataclasses-json==0.6.4
defusedxml==0.7.1
demoji==1.1.0
diff-match-patch==20230430
django-ace==1.42.0
django-admin-autocomplete-filter==0.7.1
django-admin-confirm==1.0.0
django-admin-extra-buttons==1.5.7
django-admin-inline-paginator==0.4.0
django-admin-sortable2==2.2.4
django-allauth==0.61.1
django-allow-cidr==0.7.1
django-auditlog==1.0.0
django-autocomplete-light==3.9.7
django-autogfk==0.5.2
django-celery-beat==2.6.0
django-celery-results==2.5.1
django-cors-headers==4.3.1
django-debug-toolbar==4.4.2
django-extensions==3.1.5
django-filter==24.1
django-fsm==2.8.1
django-guardian==3.0.3
django-import-export==3.3.7
django-jazzmin==3.0.1
django-json-widget==1.1.1
django-jsonfield-backport==1.0.5
django-localflavor==4.0
django-map-widgets==0.5.1
django-memoize==2.3.1
django-money==3.5.3
django-multiselectfield==0.1.12
django-nested-admin==4.1.3
django-phonenumber-field==7.3.0
django-polymorphic==4.1.0
django-redis==5.4.0
django-rest-polymorphic==0.1.10
django-select2==8.4.1
django-sequences==3.0
django-solo==2.2.0
django-storages==1.14.6
django-summernote==0.8.20.0
django-taggit==6.1.0
django-timezone-field==6.1.0
django-universal-uploader-widget==0.2.4
django-upload-validator==1.1.6
django-widget-tweaks==1.5.0
Django==4.2.11
djangochannelsrestframework==1.2.0
djangorestframework-guardian==0.4.0
djangorestframework-simplejwt==5.3.1
djangorestframework==3.14.0
docxcompose==1.4.0
docxtpl==0.16.8
drf-extra-fields==3.7.0
drf-model-serializer==0.0.2
drf-orjson-renderer==1.7.3
drf-spectacular==0.27.1
drf-writable-nested==0.7.0
elevenlabs==2.14.0
entrypoints==0.4
et-xmlfile==1.1.0
evolutionapi==0.1.2
exceptiongroup==1.2.0
fastjsonschema==2.19.1
file-magic==0.4.1
filetype==1.2.0
future==1.0.0
google-api-core==2.17.1
google-api-python-client==2.122.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
google-auth==2.28.2
googleapis-common-protos==1.63.0
gunicorn==21.2.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
http-ece==1.2.0
httplib2==0.22.0
httptools==0.6.1
humanize==4.9.0
hyperframe==6.0.1
hyperlink==21.0.0
idna==3.6
inclusive-django-range-fields==0.2.3
incremental==22.10.0
inflection==0.5.1
Jinja2==3.1.3
jmespath==0.10.0
jsonschema-specifications==2023.12.1
jsonschema==4.21.1
jupyter_core==5.7.2
kombu==5.3.5
levenshtein==0.25.1
lxml==5.1.0
MarkupPy==1.14
MarkupSafe==2.1.5
marshmallow==3.21.1
martor==1.6.45
mistune==0.8.4
MonthDelta==0.9.1
msgpack==1.0.8
mypy-extensions==1.0.0
nbconvert==5.6.1
nbformat==5.10.2
numpy==2.2.6
oauthlib==3.2.2
odfpy==1.4.1
openai==1.99.9
opencv-python-headless==4.12.0.88
openpyxl==3.1.2
orjson==3.11.0
packaging==24.0
pandas==2.3.2
pandocfilters==1.5.1
phonenumbers==8.13.32
pillow==10.2.0
platformdirs==4.2.0
priority==1.3.0
prompt-toolkit==3.0.43
protobuf==4.25.3
psycopg2-binary==2.9.6
py-vapid==1.9.0
pyasn1-modules==0.3.0
pyasn1==0.5.1
pybase62==0.3.2
pycparser==2.21
pycryptodome==3.13.0
pycryptodomex==3.20.0
pycurl==7.43.0.5
pydot==2.0.0
pydub==0.25.1
Pygments==2.17.2
pygraphviz==1.11
PyJWT==2.8.0
PyMuPDF==1.26.4
pyOpenSSL==25.0.0
pyparsing==3.1.2
pytesseract==0.3.13
python-crontab==3.0.0
python-dateutil==2.9.0.post0
python-docx==1.1.0
python-dotenv==1.0.1
python-magic==0.4.27
python-monkey-business==1.0.0
python-stdnum==1.19
pytz==2024.1
pyvips==2.2.2
pywebpush==1.14.1
PyYAML==6.0.1
pyzipper==0.3.6
qrcode==8.2
redis==5.0.1
referencing==0.33.0
replace_accents==0.0.5
requests-oauthlib==1.4.0
requests==2.31.0
rpds-py==0.18.0
rsa==4.9
rule-engine==4.5.3
s3transfer==0.13.1
selenium==4.21.0
sentry-sdk==1.0.0
service-identity==24.1.0
six==1.16.0
sniffio==1.3.1
sqlparse==0.4.4
StrEnum==0.4.15
tablib==3.5.0
testpath==0.6.0
tiktoken==0.11.0
traitlets==5.14.2
tweepy==4.14.0
Twisted==24.3.0
txaio==23.1.1
typing-inspect==0.9.0
tzdata==2024.1
tzlocal==5.2
unidecode==1.3.8
Unipath==1.1
uritemplate==4.1.1
urllib3>=1.25.4,<1.27
uvicorn==0.28.0
uvloop==0.19.0
vine==5.1.0
vobject==0.9.6.1
watchfiles==0.21.0
wcwidth==0.2.13
webdriver-manager==4.0.1
webencodings==0.5.1
websockets==12.0
Werkzeug==0.14.1
whitenoise==6.6.0
xhtml2pdf==0.2.17
xlrd==2.0.1
XlsxWriter==3.2.0
xlwt==1.3.0
zope.interface==6.2
faker==19.6.2